Edit category

+++ server/admin/admin.editor.controller.ts
@@ -22,38 +22,41 @@ export const AdminCreateEditor = async (
       return res.status(400).json({ error: "Error creating editor." });
     }
 
-    res.status(200).json(
-      {
-        name: newEditor.name,
-        email: newEditor.email,
-        status: newEditor.status,
-        role: newEditor.role,
-        createdAt: newEditor.createdAt,
-      }
-    );
+    res.status(200).json({
+      name: newEditor.name,
+      email: newEditor.email,
+      status: newEditor.status,
+      role: newEditor.role,
+      createdAt: newEditor.createdAt,
+    });
   } catch (error) {
     console.log("Create editor", error);
     return res.status(500).json("Internal server error");
   }
 };
 
-export const AdminUpdateEditor = async (req:ExtendedRequest,res:Response) =>{
-
-
-}
-
-export const AdminDeleteEditor = async(req:ExtendedRequest,res:Response) => {
-    try{
-        const {id} =req.params;
-
-        const user = await DeleteUser(Number(id))
-
-        if(!user){
-            return res.status(404).json({ error: "Not found User" });
-        }
-        res.json({error: null,user})
-    }catch(error){
-        console.log("Delete user", error);
-    return res.status(500).json("Internal server error");
+export const AdminDeleteEditor = async (
+  req: ExtendedRequest,
+  res: Response,
+) => {
+  try {
+    if (!req.user) {
+      return res.status(201).json({ error: "Acess denied" });
     }
-}
+    const { id } = req.params;
+    if (req.user.id === Number(id)) {
+      return res
+        .status(403)
+        .json({ error: "You can't exclude yourself." });
+    }
+    const user = await DeleteUser(Number(id));
+
+    if (!user) {
+      return res.status(404).json({ error: "Not found User" });
+    }
+    res.json({ error: null, user });
+  } catch (error) {
+    console.log("Delete user", error);
+    return res.status(500).json("Internal server error");
+  }
+};
